+++ tomcat/src/main/java/nextstep/jwp/controller/LoginFailedController.java
@@ -1,19 +1,22 @@
-package org.apache.coyote.http11;
+package nextstep.jwp.controller;
 
+import org.apache.coyote.http11.request.Extension;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.apache.coyote.http11.response.HttpStatus;
 import org.apache.support.ResourceFindUtils;
 
-public class NotFoundController implements Controller{
+public class LoginFailedController implements Controller {
+
+    private static final String PATH = "/401.html";
 
     @Override
     public HttpResponse service(HttpRequest request) {
-        final String responseBody = ResourceFindUtils.getResourceFile("/404.html");
+        final String responseBody = ResourceFindUtils.getResourceFile(PATH);
+
         return new HttpResponse.Builder()
                 .status(HttpStatus.FOUND)
-                .contentType("text/html")
-                .location("/404.html")
+                .contentType(Extension.HTML.getContentType())
                 .responseBody(responseBody)
                 .build();
     }
