+++ tomcat/src/main/java/nextstep/jwp/controller/RegisterController.java
@@ -1,19 +1,25 @@
-package org.apache.coyote.http11;
+package nextstep.jwp.controller;
 
 import nextstep.jwp.exception.UncheckedServletException;
+import org.apache.coyote.http11.request.Extension;
 import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.request.Path;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.apache.coyote.http11.response.HttpStatus;
+import org.apache.support.ResourceFindUtils;
 
-public class HomeController implements Controller{
+public class RegisterController implements Controller {
 
     @Override
     public HttpResponse service(HttpRequest request) {
+        final Path path = request.getPath();
         if (request.isGetMethod()) {
+            final String responseBody = ResourceFindUtils
+                    .getResourceFile(path.getResource() + Extension.HTML.getExtension());
             return new HttpResponse.Builder()
                     .status(HttpStatus.OK)
-                    .contentType("text/html")
-                    .responseBody("Hello world!")
+                    .contentType(path.getContentType())
+                    .responseBody(responseBody)
                     .build();
         }
