+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -11,6 +11,7 @@ import nextstep.jwp.exception.UncheckedServletException;
 import org.apache.coyote.Processor;
 import org.apache.coyote.http11.exception.ResourceNotFoundException;
 import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.response.ErrorResponse;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
