+++ tomcat/src/main/java/org/apache/coyote/http11/response/ErrorResponse.java
@@ -1,7 +1,4 @@
-package org.apache.coyote.http11;
-
-import org.apache.coyote.http11.response.HttpResponse;
-import org.apache.coyote.http11.response.HttpStatus;
+package org.apache.coyote.http11.response;
 
 public class ErrorResponse {
